RedbubbleScraper/requirements.txt
requests==2.32.3
beautifulsoup4==4.13.4
pandas==2.2.3
python-dotenv==1.1.0
fake-useragent==2.2.0
lxml==5.1.0 